requirements.txt
ccxt
python-telegram-bot==20.7
requests
